create ConsumerTrace and ProducerTrace
